Move module-specific services to their modules, and add support for module setups instead to run the module service setups
