+++ src/events/ClientReady.ts
@@ -1,8 +1,7 @@
 import { Events } from 'discord.js';
 import { registerCommands } from '../services/CommandRegisterer';
 import { updateGuilds } from '../services/GuildUpdater';
-import { startReminderService } from '../services/ReminderService';
-import { syncAllEnabledGuilds as syncReactionRoles } from '../modules/reactionRoles/ReactionRolesService';
+import { runModuleSetups } from '../modules';
 
 export const name = Events.ClientReady;
 export const once = true;
@@ -20,7 +19,5 @@ export async function execute(client) {
     await updateGuilds(guilds);
 
     void registerCommands(guilds);
-    void syncReactionRoles();
-    
-    void startReminderService();
+    void runModuleSetups();
 }
