+++ src/modules/levels/commands/Level.ts
@@ -1,6 +1,6 @@
 import { CommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { getCommandUserGuild } from '../../../services/CommandUserGuildResolver';
-import { calculateLevel } from '../../../services/LevelCalculator';
+import { calculateLevel } from '../LevelCalculator';
 
 export const data = new SlashCommandBuilder()
     .setName('level')
