+++ src/modules/reactionRoles/index.ts
@@ -1,5 +1,6 @@
 import * as reactionAdd from './handlers/ReactionAdd';
 import * as reactionRemove from './handlers/ReactionRemove';
+import { syncAllEnabledGuilds } from './ReactionRolesService';
 
 export const name = 'reactionRoles';
 
@@ -7,3 +8,7 @@ export const eventHandlers = [
     reactionAdd,
     reactionRemove
 ];
+
+export async function setup() {
+    await syncAllEnabledGuilds();
+}
