+++ src/modules/remindme/index.ts
@@ -1,7 +1,12 @@
 import * as remindme from './commands/Remindme';
+import { startReminderService } from './ReminderService';
 
 export const name = 'remindme';
 
 export const commands = {
     remindme,
+}
+
+export async function setup() {
+    startReminderService();
 }
